requirements.txt
pdf2image>=1.16.3
Pillow>=10.0.0
prompt_toolkit>=3.0.0